The segmentation preview approach was copying the entire dataset before cropping to the ROI.  Although this is okay for small volumes, when testing largeer volumes this becomes an excessive memory burden.  The code was rewritten such that MicroBlockObjectSetup tests for ROI and crops the image.  The SegementationPreviewer thus only receives a small image that referenced from the crop.
