+++ pom.xml
@@ -35,7 +35,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
     <version>1.0alpha-SNAPSHOT</version>
 
 
-    <name>VTEA_truth</name>
+    <name>VTEA master</name>
     <description>Volumetric analysis</description>
     
     <properties>
